new nuiSliderAttribute
git-svn-id: svn://libnui.net/nui/trunk/nui3@4852 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ include/nuiWidgetMatcher.h
@@ -119,7 +119,7 @@ public:
 
   bool Match(nuiWidget*& pWidget)
   {
-    int32 i = mpCondition.size() - 1;
+    int32 i = (int32)mpCondition.size() - 1;
     while (i >= 0)
     {
       if (!MatchOne(mpCondition[i], pWidget))
